Updating Sales order migration

+++ database/migrations/sales_order_battery/2024_06_03_150323_update_price_net_in_sales_order_battery_table.php
@@ -14,7 +14,7 @@ class UpdatePriceNetInSalesOrderBatteryTable extends Migration
     public function up()
     {
         Schema::table('sales_order_battery', function (Blueprint $table) {
-            $table->double('price_net')->default(0)->change();
+            $table->float('price_net')->default(0)->change();
         });
     }
